Major refactor to optimization DAG to make use of WebdavFileSystem. New day report DAG to visualize day results validation with experimental data

+++ optimization/src/solhycool_optimization/utils/serialization.py
@@ -1,3 +1,4 @@
+from typing import Literal
 import shutil
 import tarfile
 from enum import Enum
@@ -73,7 +74,7 @@ def get_fitness_history(algo_id: str, algo_log: pg.algorithm | pd.DataFrame |  l
     
     return fitness_history
 
-def load_optimization_results(results_path: Path, ) -> tuple[pd.DataFrame, str]:
+def load_optimization_results(results_path: Path, eval_time_from_parent_fn: bool = False) -> tuple[pd.DataFrame, str]:
     """    Load optimization results from a compressed archive and return the results DataFrame.
     Args:
         results_path (Path): Path to the compressed optimization results archive.
@@ -88,38 +89,72 @@ def load_optimization_results(results_path: Path, ) -> tuple[pd.DataFrame, str]:
 
     from solhycool_optimization import DayResults # Import here to avoid circular import issues
 
-    # Define a temporary extraction directory
-    extract_dir = results_path.parent / "extracted_results"
-    extract_dir.mkdir(exist_ok=True)
+    if results_path.is_dir():
+        # If results_path is a directory, assume it contains the results.h5 file directly
+        h5_files = list(results_path.glob("*.h5"))
+        if not h5_files:
+            raise FileNotFoundError(f"No results with .h5 file format found in directory {results_path}")
+        if len(h5_files) > 1:
+            raise ValueError(f"Multiple .h5 files found in directory {results_path}. Please ensure only one results.h5 file is present.")
+        
+        results_h5 = h5_files[0]
+        
+        # Load with DayResults.initialize
+        day_results = DayResults.initialize(results_h5)
+            
+    else:
+        # Define a temporary extraction directory
+        extract_dir = results_path.parent / "extracted_results"
+        extract_dir.mkdir(exist_ok=True)
 
-    # Extract the tar.gz archive
-    with tarfile.open(results_path, "r:gz") as tar:
-        tar.extractall(path=extract_dir, filter="data")
+        # Extract the tar.gz archive
+        with tarfile.open(results_path, "r:gz") as tar:
+            tar.extractall(path=extract_dir, filter="data")
 
-    # Find the results.h5 file (assuming only one)
-    results_h5 = next(extract_dir.rglob("results.h5"))
+        # Find the results.h5 file (assuming only one)
+        results_h5 = next(extract_dir.rglob("results.h5"))
 
-    # Load with DayResults.initialize
-    day_results = DayResults.initialize(results_h5)
+        # Load with DayResults.initialize
+        day_results = DayResults.initialize(results_h5)
 
-    # Optional: clean up extracted files after loading (uncomment to enable)
-    shutil.rmtree(extract_dir)
+        # Optional: clean up extracted files after loading (uncomment to enable)
+        shutil.rmtree(extract_dir)
     
     # Extract date from filename
-    eval_time = results_path.stem.split("eval_at_")[-1].split(".")[0]  
-    
+    if day_results.eval_at is None or eval_time_from_parent_fn:
+        eval_time = results_path.stem.split("eval_at_")[-1].split(".")[0]
+    else:
+        eval_time = day_results.eval_at
+                
     return day_results.df_results, eval_time
 
-def load_multiple_optimization_results(results_dir: Path, date_str: str) -> tuple[list[pd.DataFrame], list[str], pd.DataFrame]:
+def load_multiple_optimization_results(
+    results_dir: Path, 
+    date_str: str,
+    look_for: Literal["dir", "tar.gz"] = "tar.gz",
+    eval_time_from_parent_fn: bool = False
+) -> tuple[list[pd.DataFrame], list[str], pd.DataFrame]:
     """
         The joined dataframe is built by "masking by precedence" behavior, where:
             - Each DataFrame contains indexed values (e.g., by step, timestamp, or some id).
             - If a newer DataFrame contains values at a given index, you want to ignore all values from older DataFrames starting from that index onward, even for rows not duplicated.
             - This means older DataFrames contribute only up to the earliest index present in any newer DataFrame.
     """
-    results_files = sorted(results_dir.glob(f"*{date_str}*.tar.gz"))  # Oldest to newest
     
-    dfs, eval_times = zip(*[load_optimization_results(results_path) for results_path in results_files])
+    # print(f"{look_for=}, {date_str=}, {results_dir=}")
+    
+    if look_for == "dir":
+        results_files = sorted(results_dir.glob(f"*{date_str}*"))
+    else:
+        results_files = sorted(results_dir.glob(f"*{date_str}*.{look_for}"))  # Oldest to newest
+        
+    if len(results_files) == 0:
+        raise FileNotFoundError(f"No results found in directory {results_dir} for date {date_str} with look_for={look_for}")
+
+    dfs, eval_times = zip(*[
+        load_optimization_results(results_path, eval_time_from_parent_fn) 
+        for results_path in results_files
+    ])
     
     # Start merging from oldest to newest, but apply masking
     kept_indices = set()
